feat: sell

+++ Assets/Economic/Interfaces/IWallet.cs
@@ -4,12 +4,12 @@ namespace Factory
 {
 	public interface IWallet
 	{
-		public long GetBalance(CurrencyType type);
-		public void CurrencyTransfer(CurrencyType type, long value);
-		public void CurrencyTransfer(CurrencyType type, long value, out long walletBalance);
-		public void CurrencySet(CurrencyType type, long value);
+		public double GetBalance(CurrencyType type);
+		public void CurrencyTransfer(CurrencyType type, double value);
+		public void CurrencyTransfer(CurrencyType type, double value, out double walletBalance);
+		public void CurrencySet(CurrencyType type, double value);
 		public void CurrencySetDefault(CurrencyType type);
-		public void CurrencySetAll(long value);
+		public void CurrencySetAll(double value);
 		public void CurrencySetDefaultAll();
 	}
 }
