+++ Assets/Product/Scripts/ProductTemplate.cs
@@ -5,7 +5,7 @@ namespace Factory
     public abstract class ProductTemplate : ScriptableObject
 	{
         public Product Product;
-        public abstract long Price(int level);
+        public abstract double Price(int level);
         public abstract int MaxLevel();
     }
 }
